+++ exchange/stable_ex.go
@@ -19,7 +19,8 @@ func (self *StableEx) TokenAddresses() (map[string]ethereum.Address, error) {
 	// returning admin multisig. In case anyone sent dgx to this address,
 	// we can still get it.
 	return map[string]ethereum.Address{
-		"DGX": ethereum.HexToAddress("0xFDF28Bf25779ED4cA74e958d54653260af604C20"),
+		"DGX":  ethereum.HexToAddress("0xFDF28Bf25779ED4cA74e958d54653260af604C20"),
+		"WBTC": ethereum.HexToAddress("0xFDF28Bf25779ED4cA74e958d54653260af604C20"),
 	}, nil
 }